SDKs/requirements.txt
google-genai
python-dotenv
fpdf
matplotlib
yfinance
pandas

